Extract Dashboard to its own class
