+++ src/models/CampaignDb.ts
@@ -9,6 +9,7 @@ import { Maybe, MetadataResponse, DataEntryForm, Section } from "./db.types";
 import { Metadata, DataSet, Response } from "./db.types";
 import { getDaysRange, toISOStringNoTZ } from "../utils/date";
 import { getDataElements } from "./AntigensDisaggregation";
+import { Dashboard } from "./Dashboard";
 
 interface DataSetWithSections {
     sections: Array<{ id: string; name: string; dataSet: { id: string } }>;
@@ -41,7 +42,8 @@ export default class CampaignDb {
         }
         const startDate = moment(campaign.startDate).startOf("day");
         const endDate = moment(campaign.endDate).endOf("day");
-        const { dashboard, charts, reportTables } = await db.createDashboard(
+        const blankDashboard = Dashboard.build(db);
+        const { dashboard, charts, reportTables } = await blankDashboard.createDashboard(
             campaign.name,
             campaign.organisationUnits,
             campaign.antigens,
